not fix test

+++ etl-service/src/main/java/com/sitefilm/etl/dto/core/person/PersonCrewDto.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Builder
 public class PersonCrewDto {
         @JsonProperty("id")
-        private Long externalId;
+        private Integer externalId;
         @JsonProperty("department")
         String department;
         @JsonProperty("job")
